+++ arch/x86/64/source/ArchThreads.cpp
@@ -106,7 +106,7 @@ size_t ArchThreads::testSetLock(size_t &lock, size_t new_value)
   return __sync_lock_test_and_set(&lock,new_value);
 }
 
-uint32 ArchThreads::atomic_add(uint32 &value, int32 increment)
+uint64 ArchThreads::atomic_add(uint64 &value, int64 increment)
 {
   int32 ret=increment;
   __asm__ __volatile__(
@@ -117,11 +117,10 @@ uint32 ArchThreads::atomic_add(uint32 &value, int32 increment)
   return ret;
 }
 
-int32 ArchThreads::atomic_add(int32 &value, int32 increment)
+int64 ArchThreads::atomic_add(int64 &value, int64 increment)
 {
-  return (int32) ArchThreads::atomic_add((uint32 &) value, increment);
+  return (int64) ArchThreads::atomic_add((uint64 &) value, increment);
 }
-
 
 void ArchThreads::printThreadRegisters(Thread *thread, bool verbose)
 {
